restructure tests dir
