+++ packages/sdk/src/test/config/index.ts
@@ -1,0 +1,1 @@
+export * from './base'
