+++ packages/sdk/src/test/constants/regexes.ts
@@ -1,0 +1,1 @@
+export const ADDRESS_REGEX = /^0x[a-fA-F0-9]{40}/
